use the actual Android names for app users
This provides a transparent mixing of Debian and Android, and as long as
this package is installed as part of the Lil' Debi install process, then
this package will create all of the conflicting group names with the
Android uid, and the Debian packages should work just fine with those uids.
For example, Debian also has a bluetooth group.