force all groups to use Android gid (Closes: #767677)
Since this package is only meant to be installed as early as possible on a
Debian chroot running on Android, this forces the gids to match Android's
gids.  That ensures that things are properly integrated.

This also allows piuparts to test this package, or informs users of issues
if they happen to install this package on Android-based chroot after many
other packages have been installed.